Adjusted Adapter Symmetry tests, some new failures
