+++ johnzon-jsonb/src/test/java/org/apache/johnzon/jsonb/symmetry/adapter/map/MapAdapterOnClassSimpleTest.java
@@ -29,7 +29,7 @@ public class MapAdapterOnClassSimpleTest extends MapAdapterOnClassTest {
 
     @Override
     public void assertRead(final Jsonb jsonb) {
-        final String json = "{\"email\":\"test@domain.com\"}";
+        final String json = "{\"email\":{\"user\":\"test\",\"domain\":\"domain.com\"}}";
         final MapAdapterPrecedenceConfigClassTest.Contact actual = jsonb.fromJson(json, MapAdapterPrecedenceConfigClassTest.Contact.class);
         assertEquals("Contact{email=test@domain.com:EmailClass.adaptFromJson}", actual.toString());
         assertEquals("Contact.<init>\n" +
@@ -45,7 +45,7 @@ public class MapAdapterOnClassSimpleTest extends MapAdapterOnClassTest {
         reset();
 
         final String json = jsonb.toJson(contact);
-        assertEquals("{\"email\":\"test@domain.com:EmailClass.adaptToJson\"}", json);
+        assertEquals("{\"email\":{\"user\":\"test\",\"domain\":\"domain.com\",\"call\":\"EmailClass.adaptToJson\"}}", json);
         assertEquals("Contact.getEmail\n" +
                 "EmailClass.adaptToJson", calls());
     }
